refactor: extract CLI commands into individual modules with thin router
Split the monolithic cli.ts (~1440 lines) into focused command modules
under src/commands/, leaving cli.ts as a thin router (~180 lines).

Command modules created:
- commands/shared.ts — Options, parseOptions, user config, withClient,
  model auto-selection, thread start/resume, result printing, PID mgmt
- commands/run.ts — run + resume handler
- commands/review.ts — review (all modes + resume) handler
- commands/threads.ts — threads list, output, progress, delete, clean
- commands/kill.ts — kill handler
- commands/config.ts — config get/set, models, health
- commands/approve.ts — approve + decline handlers

cli.ts now handles signal registration, help text, command extraction,
data directory setup, and dynamic dispatch to command modules. The
deprecated 'jobs' command routes to threads with a deprecation warning.
Added 'threads' as the canonical command name.
